+++ src/vm/mod.rs
@@ -3,5 +3,6 @@ pub mod jsvalue;
 pub mod codegen;
 pub mod constant;
 pub mod error;
+pub mod factory;
 pub mod frame;
 pub mod vm;
